requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
pydantic==2.11.7
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
python-multipart==0.0.9
